+++ _examples/http-server/listen-addr-public/main.go
@@ -28,6 +28,7 @@ func main() {
 					{
 						Name: "MyApp",
 						Addr: ":8080",
+						Hostname: "your-custom-sub-domain.ngrok.io", // optionally
 					},
 				},
 			},